% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONARNetCDF4-package.R
\docType{package}
\name{SONARNetCDF4}
\alias{SONARNetCDF4}
\alias{SONARNetCDF4-package}
\title{Title of the package}
\description{
Description of the package
}
\details{
Details of the package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnejohannesholmin/SONARNetCDF4}
  \item Report bugs at \url{https://github.com/arnejohannesholmin/SONARNetCDF4/issues}
}

}
\author{
\strong{Maintainer}: Given Name Family Name \email{arnejh@hi.no}

}
