% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONARNetCDF4-package.R
\docType{package}
\name{sonaRnetcdf4}
\alias{sonaRnetcdf4}
\alias{sonaRnetcdf4-package}
\title{Tools to Read and Write netCDF4 Files in the SONAR-netCDF4 Convension}
\description{
Functions to read and write netCDF4 files in the SONAR-netCDF4 convension, which is a standard for SONAR and echosounder data and derived data defined by the International Council for the Exploration of the Sea (ICES).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnejohannesholmin/sonaRnetcdf4}
  \item Report bugs at \url{https://github.com/arnejohannesholmin/sonaRnetcdf4/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

Other contributors:
\itemize{
  \item Norwegian Institute of Marine Research [copyright holder, funder]
}

}
